third_party/blink/web_tests/fast/forms/select/select-selectedIndex-noAnchorIndex-crash-expected.txt

PASS multiselect.selectedIndex is -1
PASS successfullyParsed is true

TEST COMPLETE

item a
item b
